# Use native code for validating VAAs