@thmzlt 2023-05-19
> do we have enough logging to catch any bugs in the logic quickly? (e.g., do we log on an invalid VAA)?

Yes. It [logs](https://github.com/pyth-network/pyth-crosschain/blob/main/price_service/server/src/listen.ts#L318) and also causes the processing of a VAA to halt (i.e. it will manifest in the metrics).